feat(dashboard): rebuild the project card around its last deployment
Replaces the app icon stack with the latest deployment across the
project's apps: commit message, branch, age, the real deployment status
badge, and a count line for apps, keyspaces and ratelimits. Projects
with nothing deployed say so and stop there.

The deployment generator moves into the prototype folder so the list
page can use it; the overview path re-exports it.
