+++ web/apps/dashboard/app/(app)/[workspaceSlug]/projects/_components/prototype/projects-prototype.tsx
@@ -16,6 +16,7 @@ import {
   DebugCommand,
   useAgentDismissed,
   useAgentStyle,
+  useChartScheme,
   useMark,
   useRowVariant,
   useScenario,
@@ -35,6 +36,7 @@ function ProjectsPrototypeInner() {
   const { scenario, setScenario } = useScenario();
   const { variant, setVariant } = useRowVariant();
   const { mark, setMark } = useMark();
+  const { chartScheme, setChartScheme } = useChartScheme();
   const { agentStyle, setAgentStyle } = useAgentStyle();
   const [agentDismissed, setAgentDismissed] = useAgentDismissed();
   const { worlds, resetWorlds } = usePrototypeWorlds();
@@ -59,7 +61,12 @@ function ProjectsPrototypeInner() {
                 <CreateProjectButton workspaceSlug={workspace.slug} variant="outline" />
               </ProjectsEmptyCard>
             ) : (
-              <ProjectGrid projects={world.projects} workspaceSlug={workspace.slug} />
+              <ProjectGrid
+                projects={world.projects}
+                keyspaces={world.keyspaces}
+                ratelimits={world.ratelimits}
+                workspaceSlug={workspace.slug}
+              />
             )}
           </div>
           <Rail
@@ -84,6 +91,8 @@ function ProjectsPrototypeInner() {
         onAgentStyle={setAgentStyle}
         agentDismissed={agentDismissed}
         onToggleAgent={() => setAgentDismissed(!agentDismissed)}
+        chartScheme={chartScheme}
+        onChartScheme={setChartScheme}
         onReset={resetWorlds}
       />
     </PageContainer>
